Implemented animal breeding
Players can now breed animals of the same type and different gender to try to create babies.

Made a method to display all animals with an index in the Player class.

Removed prompts to check if you entered the correct name or chose the correct animal to sell because they were annoying to have to answer all the time.